[Change] Extract drive googleapis utility functions to its own file for reusability
